pyproject.toml
[project]
name = "github_connector"
version = "0.1.0"
description = "Library for fetching repository data from github"
authors = [
    {name = "Ekene Agwu"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)"
]
